Updated DownloadCommand to be an integration test
